Fix driver control and lights confusion
The issue was that the driver control indicator and the lights indicator were not
in the same format as the rest of the dashboard control indicators, and it confused
people that looked at it; the lights label and the driver control label were on the same
line and it looked like they were the same indicator, and they didn't have the on
label to indicate whether the driver control was on or the lights was on.
The commit fixes the issue by seperating the driver control and lights into seperate
entities and added on labels.