SFT-88 Modified build output structure to allow shadow builds
Remove explicit references to a `build` folder, so the outputs can be placed anywhere the developer wants. The proper way of using qmake is to create the build directory, and calling qmake from that build directory to generate the makefile.

The executable will now be located in `<build-folder>/bin`. Other libraries will have a folder containing `.lib`,`.moc`, `.obj` folders for those artifacts.

In addition, a new copyconfigfile command was added, that will copy the config.ini next to the executable post-build.